+++ src/New/Entity/AttachEffectClass.cpp
@@ -867,7 +867,7 @@ int AttachEffectClass::RemoveAllOfType(AttachEffectTypeClass* pType, TechnoClass
 
 	auto const targetAEs = &pTargetExt->AttachedEffects;
 	std::vector<std::unique_ptr<AttachEffectClass>>::iterator it;
-	std::vector<std::pair<WeaponTypeClass*, TechnoClass*>> expireWeapons;
+	std::vector<ExpireWeaponData> expireWeapons;
 
 	for (it = targetAEs->begin(); it != targetAEs->end(); )
 	{
@@ -887,12 +887,14 @@ int AttachEffectClass::RemoveAllOfType(AttachEffectTypeClass* pType, TechnoClass
 				{
 					if (pType->ExpireWeapon_UseInvokerAsOwner)
 					{
-						if (auto const pInvoker = attachEffect->Invoker)
-							expireWeapons.push_back(std::make_pair(pType->ExpireWeapon, pInvoker));
+						const auto pInvoker = attachEffect->Invoker;
+
+						if (pInvoker || !pType->ExpireWeapon_InvokerMustAlive)
+							expireWeapons.emplace_back(pType->ExpireWeapon, pInvoker, attachEffect->InvokerHouse);
 					}
 					else
 					{
-						expireWeapons.push_back(std::make_pair(pType->ExpireWeapon, pTarget));
+						expireWeapons.emplace_back(pType->ExpireWeapon, pTarget, pTarget->Owner);
 					}
 				}
 			}
@@ -921,11 +923,8 @@ int AttachEffectClass::RemoveAllOfType(AttachEffectTypeClass* pType, TechnoClass
 
 	auto const coords = pTarget->GetCoords();
 
-	for (auto const& pair : expireWeapons)
-	{
-		auto const pInvoker = pair.second;
-		WeaponTypeExt::DetonateAt(pair.first, coords, pInvoker, pInvoker->Owner, pTarget);
-	}
+	for (auto const& [pWeapon, pInvoker, pInvokerHouse] : expireWeapons)
+		WeaponTypeExt::DetonateAt(pWeapon, coords, pInvoker, pInvokerHouse, pTarget);
 
 	return detachedCount;
 }